fix(linux): ensure icon works in sandboxed environments (#43)
* refactor: only pass icon as base64 for macOS

Only macOS requires the icon to be passed as a base64 string. No need
to do this for other platforms.

* fix(linux): ensure icon works in sandboxed environments

When running in a sandboxed environment, such as a Flatpak or Snap, the
icon should be passed as the icon name, not a path.

This is required because when running in a sandbox, paths are not the
same as seen by the app and the host system.